[YS-616] refactor: 공고 둘러보기 컴포넌트 위치 수정
